requirements.txt
crewai
crewai-tools
langchain==0.1.20
langchain-groq==0.1.3
langchain-community==0.0.38
langsmith==0.1.60
fastapi==0.111.0
uvicorn==0.29.0
streamlit==1.35.0
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.7.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.128.0
PyGithub==2.3.0
httpx==0.27.0
rich==13.7.1
loguru==0.7.2
pytest==8.2.0
pytest-asyncio==0.23.7